+++ src/main/java/ru/liga/prerevolutiontelegrammtinderbot/telegram_tinder_pre_bot/botapi/handlers/fillingProfile/AskNameHandler.java
@@ -25,7 +25,7 @@ public class AskNameHandler implements InputMessageHandler {
 
     @Override
     public BotApiMethod<?> handleUpdate(Update update) {
-        int userID = Math.toIntExact(UpdateHandler.getId(update));
+        long userID = UpdateHandler.getId(update);
 
         String chatId = UpdateHandler.getChatId(update);
